Override `<button>` padding from page styles in toolbar
Revert to hard-coded padding for the toolbar buttons to override any
default padding the page may have for `<button>` elements. This affects
PDF.js amongst other apps/sites. Instead of setting the padding to 0 as
before, set it to `1px 6px` which matches Chrome's defaults that looked
reasonable.

A better solution will be to use Shadow DOM for the whole sidebar UI,
but that is a larger change.