Try to fetch all service-specified groups in a single API call
Rework fetching of groups specified in the client's settings to first
fetch all the groups that the user is a member of and then only make
separate API calls to fetch groups specified in the configuration which
are not in that list.

In the case of the LMS app, which is the main use case for this setting,
this reduces the load on the backend by only making one request to fetch
an LMS user's groups rather than one per course section.

In the process the existing parameterized test was split into separate
tests since the logic for each case was sufficiently different that it
made more sense.