When multiple unsaved annotations exist, scroll to the newest
It is possible for multiple unsaved annotations to exist if a user
creates a new annotation, enters some text but does not save, and then
creates another new annotation. In this case, make the thread list
always scroll to the newest. A side effect of the implementation is that
if this second annotation is then deleted or saved, the thread list will
scroll back to the previous unsaved annotation. This behavior seems
useful, so I've kept it.