Prevent scrolling to annotations from links or buttons
If an event is triggered from a link or button or a descendent of a link or button, then prevent scrollToAnnotation from firing. It is assumed in these cases that a more specific action was intended and the page shall not scroll to the underlying annotation. Additional, dom-element.js has moved to /shared since its now used in both the annotator and sidebar