Typecheck `src/annotator/{config, util}` and `src/sidebar/*.js`
Enable typechecking for the following directories:

 - src/annotator/config
 - src/annotator/util
 - src/sidebar (excluding `ga.js`, `icons.js` and `index.js`)

Many functions in these directories are still not annotated, but this
will be added over time. `ga.js` was excluded because it deals with
globals added by Google Analytics. `icons.js` was excluded because TS
doesn't know how to handle `require('.../some-icon.svg')`. `index.js`
was excluded because that imports (and thus triggers checking for) every
other file in the sidebar app. It will need to be enabled last.