Typecheck SvgIcon
Typecheck `shared/components/svg-icon.js`. This enables edit-time rather than
runtime detection of incorrect usage, as well as enabling UI component
props to reference types (eg. API objects) defined elsewhere with JSDoc.

Props are currently declared with both JSDoc and via
`SvgIcon.propTypes`. The intent is that the JSDoc is the canonical
definition, but `SvgIcon.propTypes` is still useful to catch incorrect
usage from other parts of the code which are currently not typechecked.