Typecheck UI components in src/annotator/components
This follows the same pattern that was introduced for `SvgIcon` in
`propTypes` declarations are kept around for use at runtime by other
code which is not yet typechecked.

An issue that I encountered is that TS doesn't know about the custom
HTML element types we use. For the moment I have ignored these with
`@ts-ignore`, although I believe it should be possible to teach it.