Work around a button rendering issue in Safari
Work around an issue where the child of a `<button>` element with
`height: 100%` does not fill the button in Safari in certain
circumstances, but does in Firefox and Chrome. See [1] for a reduced
test case.

This caused the drop-down menu toggle button on the right-hand side of the
"Post to <group name>" control to be squashed.

The workaround here is to use a `<div role="button">` instead.

Fixes https://github.com/hypothesis/client/issues/2302

[1] https://github.com/hypothesis/client/issues/2302#issuecomment-654632234