Flush timers after each test
If a test sets up a timer that expires some time after the test
completes, this can cause confusing "failures" in other tests. It can
also cause potential errors to be missed if the timer would trigger an
error when it fires, but the timeout never actually gets to fire because
the test suite completes first.

Currently flushing is only enabled for the two most common timer
functions that our code and libraries use, but we may need to add others
(eg. `requestIdleCallback`, `setInterval`) in future.