Improve typechecking for Stores
Add SidebarStore type. This is an aggregation of the base ReduxStore + each one of our stores baked in. The fist store added here is simply the Activity store (as a first example)

There are some assumptions when making store types:

1. It is assumed they are transformed in `createStore` where the selectors and actions are flattened.

2. Some meta data (namely, the init method, namespace and update method) are refactored or removed.

These transformations are implied through the typing which explicitly ignores these methods and structure of the raw store before the transformation in create-store.