Refactor store annotations tests to use public store API
The majority of store module tests now work by creating an instance of the
store with a subset of store modules and then using the store's public
methods to manipulate data.

There were some tests for `store/modules/annotations.js` that pre-date
this convention and still called functions exported from the module
directly, passing in `store.getState()`.

Refactor these tests to use the public API for consistency and also to
make future changes to the internal store structure easier.