Do not set initial empty search query
The way that the search interface is currently implemented,
clicking on the "search" icon in the top bar actually
submits the search form (with an empty query).

Ignoring empty queries if a query has not yet been set
alleviates "searching" for nothing.

This also prevents expanded threads from being re-collapsed
when clicking on the search icon.