Fix a previously silent test failure
Fix a test that was failing but failing in a way that went unreported
previously due to a combination of it occurring after the test function returned
because of a missing `await` call and because of the issue with uncaught
exceptions between tests described in https://github.com/hypothesis/client/issues/2249.

This test was failing but the failure was only reported after the test
function returned due to a missing `await`.