Replace the index page of the dev server with something more useful
Replace the index page of the dev server with a more useful landing
page. The iframe has been removed because it is something we're not
testing most of the time and there are remaining issues with the way
multiple frames are supported in the sidebar. A separate test document
can be added for this purpose in future.