Fix null-safety errors
Various methods in `DocumentMeta` assume that `this.metadata.{dc,
highwire, ...}` properties exist, but this is only true once the
`getDocumentMetadata()` function has been called.

Adjust the constructor so that these fields always exist.

Note that while this avoids an exception if calling certain methods before
`getDocumentMetadata`, it would still be a logical error if that
happened. For the moment such a mistake is just silently ignored.