Remove sidebar toggle button swipe gesture handling
Remove logic for handling swipe gestures on the sidebar's toggle button.
Trying to perform a swipe on a target as small as the toggle button is
difficult and it doesn't offer advantages over just tapping the button
to open and close. Behavior that _might_ be useful is being able to
two-finger swipe on the whole sidebar to open or close it, but this conflicts
with the browser's own swipe gesture handling. Just remove it entirely
for now.

 - Remove swipe gesture handling

 - Remove unnecessary use of jQuery

 - Ensure that the `_setupGestures` code path is exercised by tests