Suggested tag matching should not use lookback
Change the matching logic for determining whether a given suggested tag
(`tag`) matches the currently-typed tag field text (`query`).

Formerly, `tag` would match `query` if `tag` were a substring of `query`
in any position. However, this creates a lot of noise in the matches and
is not commonly what folks are after.

Now, `tag` will match `query` if:

* `tag` starts with `query` OR
* `query` occurs within `tag` after a word boundary or a non-word
   character

Thus, a `query` of "app":

* Would match the `tag` "apple"
* Would match the `tag` "crab apple" ("app" occurs after a word boundary)
* Would match the `tag` "crab.apple" ("app" occurs after non-word char)
* Would NOT match the `tag` "crabapple"