Remove editing logic from `AnnotationBody`, `Annotation`
Consolidate annotation-editing logic and sub-components within
`AnnotationEditor`. `AnnotationBody` is for rendering content, and
`Annotation` is relieved of managing editing-related callbacks.

These changes also add a new feature: When saving an annotation,
`AnnotationEditor` will check the contents of `TagEditor`'s input
element. If any text remains there that has not yet bet "committed"
as a tag, that text will be added as a tag. This is to prevent tags
accidentally getting "lost" on save.

Fixes https://github.com/hypothesis/product-backlog/issues/1122