Re-implement highlight and document event handling in Guest class
The way that the `Guest` class listened to clicks on highlights and the
document involved a DOM event-listening abstraction in `Delegator` that
is no longer used for anything else in the code. This indirection made
it more complicated to follow how these events were handled.

This commit replaces the event handling for clicks, hovers and taps on
highlights and the page with regular DOM event listeners in the `Guest`
class.