Remove unused code for listening to DOM events in `Delegator`
The `Delegator` events system was used for handling two kinds of event:

 - DOM events on highights and document elements
 - Custom events internal to the "annotator" module

The handling of DOM events has been moved into the `Guest` class, so
`Delegator` is now used only to subscribe to and publish internal
events. This commit removes the DOM event handling code.

There is an existing issue, noted in this commit, that the `destroy`
method only removed event handlers declared via the `events` property
and not those added later by `subscribe`.