Convert `Delegator` base class to JS
Convert the `Delegator` base class to JS and rewrite the documentation
to match current functionality and usage. Also simplify the `options`
initialization and add a note about why the constructor reads
`this.options` before assigning it.

I tried to convert the `Plugin` class as well but encountered an issue
with interaction between Babel-compiled derived classes and CoffeeScript
sub-classes (in particular, the `BucketBar` class), so it looks like
`Plugin` sub-classes will need to be converted first.