Add normalize and serialize converted modules
These modules contain only 1 function each and replace the 3 serialize() and normalize() functions previously found in each of the 3 coffeescript range classes. The fundamental difference is that these functions are not bound to a specific range class type and instead take a range instance param and then decided which conversion algorithm to run based on the type of instance.

This removes the circular dependencies each of the 3 range classes had and additionally groups all complex serialize and normalize logic into discrete modules.