Move highlighting logic out of Guest class
The logic for handling details of the highlights layer was spread
between the `highlighter` module and the `Guest` class.

This commit moves the remaining highlighting logic in the `Guest` class
to the `highlighter` module and rewrites it to use plain DOM APIs rather
than jQuery.

As a bonus, this makes it easy to remove the remaining direct usage of jQuery
in the `Guest` class, so that has also been done in this commit.