Remove rAF from highlight drawing and make synchronous
Bug #2502 arises from an issue with the asynchronous code in guest
that relies on `requestAnimationFrame` before drawing highlights for
a particular anchor range.

When the DOM is being mutated in multiple ways, the range present in
the synchronous `highlight` code can be different from the range once
the `requestAnimationFrame` — i.e. it could have changed again since
the range was sniffed in the `highlight` method.

The call to `range.normalize` in the rAF callback in these cases might
be operating on a range that is effectively invalid per the `normalize`
method's (unwritten) expectations. That in turn throws and then causes
a failure of highlight drawing until the document is reloaded.

Some brittleness in `range` methods and other dependencies makes this
challenging to fix easily in an asynchronous manner. It's unclear
exactly why rAF asynchronicity was added
(see ca87c1c5a5bf051e6b7dd01b4c449635796e0d51 ).

As such, we've elected for now to remove `rAF` and make `highlight`
code synchronous here.

Fixes #2502