Simplify `range.commonAncestorContainer` update
The test case given in the comment no longer fails in modern versions of Safari
and I was unable to find any open or closed bug reports mentioning
`Node.contains`. Since the comment referenced a very old version of
Safari, I think it is reasonable to assume it has long been resolved.