Fix range selector generation in Chrome and Firefox
When generating range selectors the client excludes elements from the
Hypothesis UI by filtering out elements that match the CSS selector
defined by `IGNORE_SELECTOR` in guest.js.

In a554058161 this logic was changed from jQuery's selector matching to
`element.matches(...)`. The selector used in the actual application is a
compount selector which is then inverted with `:not`:

```
:not([class^="annotator-"],[class^="hypothesis-"])
```

It turns out that the `:not(...)` pseudo-class only supports _lists_ of
selectors in Safari, but not in Chrome or Firefox [1]. Therefore the call to
`element.matches(...)` would always fail in these browsers and the
resulting annotation would only have `TextQuoteSelector` and
`TextPositionSelector` selectors but no `RangeSelector`.

This commit fixes the issue by removing the use of `:not` and instead
inverting the `element.matches(...)` test. It also adds a test case for
handling of list selectors.

[1] https://developer.mozilla.org/en-US/docs/Web/CSS/:not#Browser_compatibility