Gracefully abort startup in old browsers
Run a series of feature tests at the very start of the boot process and
abort with a console warning if they fail.

We display a console warning rather than a louder warning to avoid being
disruptive on sites maintained by publishers which embed Hypothesis and
which still want to support older browsers. On those sites the
Hypothesis annotation feature will simply be missing.