Update Babel compilation targets
Update Babel compilation targets to match the browsers we currently support,
rather than transpiling everything down to ES5.

This reduces the size of our JS bundles and improves performance by using native
implementations of ES2015+ language features. It should also avoid
occassional issues in development caused by interactions of Babel plugins that
compile native features to older syntax (eg. `transform-async-to-promises`).

The client's boot script has a different configuration than the rest of the
client because it needs to be able to run in older browsers in order to show a
warning when the browser is not supported.

The production bundle size changes resulting from this are:

Sidebar bundle: 398 KB => 329 KB
Annotator bundle: 219 KB => 169 KB