Remove checks for IE 11 in `TagEditor` component
`TagEditor` tested if the browser was IE 11 to determine whether to
check the `inputType` property of `InputEvent`. However it is not only
IE 11 that did not support this property.

From reading the comments in https://github.com/hypothesis/client/pull/1558 it
appears that the `inputType` checks was originally added to resolve an
issue with accepting suggestions by pressing Enter in an earlier version
of `TagEditor` that used the `<datalist>` element. As far as I can tell
this is no longer applicable. `<datalist>` was replaced by a custom UI
and accepting suggestions by pressing Enter works without this logic.
Therefore this commit removes it entirely.