Move `src/shared/polyfills` to `src/boot/polyfills`
The code in `polyfills/index.js` which determines which polyfills are
needed are compiled as part of the boot script and not used by other
parts of the application.

Because this code is compiled as part of the boot script it needs to be
compiled to parse in older browsers, even those which don't support the client
(eg. IE 10/11). Putting the `polyfills/` dir in side `src/boot` means
that it will inherit any custom Babel configuration for the `src/boot`
directory.