Fail more gracefully in unsupported versions of Edge Legacy
Add a test for the `Element.prepend()` method so that startup fails more
gracefully in older, unsupported versions of Edge Legacy (<= 16). This
API was chosen as a representative test because it's Chrome/Edge Legacy
support closely matches our minimum versions.

The same change was made for the LMS app in https://github.com/hypothesis/lms/pull/2209.