Move browser support list to `browserslist` key in package.json
Move the list of supported browsers from `.babelrc` to the
`browserslist` key in `package.json`. This enables other
browserslist-enabled tools such as autoprefixer to adapt their output to
the target browsers.

This reduces the size of the production CSS bundles slightly by reducing
the number of unnecessary vendor prefixed rules that are generated.

See https://github.com/postcss/autoprefixer#browsers,
https://github.com/browserslist/browserslist#queries and
https://babeljs.io/docs/en/babel-preset-env#browserslist-integration.