Add `TextPosition` and `TextRange` classes
Add a `TextRange` class which represents a region of the document as an
immutable start/end pair of (element, text position) points. Compared to
a DOM `Range`, this representation of a document region is more robust to
changes in the DOM structure of the region that don't affect the text
content, such as wrapping parts of the range in elements to insert
accessible highlights.