Don't try to resolve sidebar app URL when booting sidebar app
When attempting to create a build of the browser extension using the
local dev client and the production h service I encountered an issue where
resolving the `sidebarAppUrl` setting in `src/boot/index.js` failed.

This happens because the extension does not set this setting in the
sidebar app so the boot script fell back to trying to resolve the URL
template baked into the development client. This in turn failed
because `document.currentScript` is not set in this context.

The `sidebarAppUrl` setting is not actually needed when booting the sidebar app, so
this commit refactors the boot code to avoid generating it in this
context. This is done by moving the responsibility for determining which
part of the client to load from `boot/boot.js` into `boot/index.js` and
only resolving the necessary settings.