Convert PDF anchoring to use TextPosition / TextRange classes
The PDF anchoring code was using several different dependencies to
to handle converting between ranges and text positions, as well as
normalizing ranges to that they start and end in text nodes.

This commit simplifies the picture by converting the PDF anchoring code
to use only the `TextRange` and `TextPosition` classes for these
purposes.

 - Change the PDF anchoring code to use the `TextRange` and
   `TextPosition` classes to handle converting a DOM range into text
   offsets in `describe`

 - Change the PDF anchoring code to use `TextRange` to handle resolving
   a (start, end) pair of character offsets within a page to a `Range`

 - Improve the type information for several related functions within the
   PDF anchoring code