Clarify what happens if range is outside of text layer
Ensure that the `describe` function in `anchoring/pdf.js` fails with a
useful error if the passed range includes text outside of the PDF page's
text layer. In this case `getNodeTextLayer` will return `null` and
existing code in `describe` will throw to indicate the range starts or
ends outside of the text layer.

This will help to avoid the confusion that arose in
https://github.com/hypothesis/client/issues/1464.

This behavior is still not really useful from the end user's point of
view, but at least the error is better. Improving the actual behavior
will come later.