Refactor `LoadAnnotationsService#load` for flexibility
`load` was written with `SidebarView` in mind, assuming the presence
of both a `groupId` and one or more `uris`. This allows the loading of
all annotations that pertain to a specific group on one or more
specific document URIs.

For Notebook, we'd like to be able to, initially, load all annotations
for a `group` but for all document URIs.

Make `load` take `options` instead of individual search arguments. Leave
requirement for a `groupId` in place for now.