Remove all links, scripts and stylesheets when client is unloaded
Ensure that all links, script and stylesheet elements added to the page
by the boot script are removed when the client is unloaded.

This fixes a problem where the recently added `<link>` for the notebook
app was not removed when the page was unloaded, causing attempts to
activate the browser extension on a page where the client had previously
been loaded and then removed, to fail.

The implementation works by tagging all elements with a
`hypothesis-asset` attribute in the boot script and then removing these
elements when the client is unloaded.