Do not select/focus the input element on iOS
The text in the input element of the `annotation-share-control` is
selected and focused when the element is created. This is to
facilitate copy/paste via keyboard shorcuts. This selection/focus mechanism is working well except for iOS (both Chrome and Safari) where:

* the selection/focus is not shown in the input field except by clicking on the input element.

* the selection overflows the right margin of the input field

I tested `overflow: hidden` on the input element but that causes the selection to overflow on the left margin

Because of these two issues, I recommend to supress selection/focus on iOS devices.